Added timelock contract, Interfaces for timelock and panacoint to be used in governer. Updated governer to add more functionality

+++ governance/contracts/GovernerCore.sol
@@ -94,7 +94,8 @@ contract GovernerCore {
         uint8 support;
 
         // @notice The number of votes the voter had, which were cast
-        uint96 votes;
+        uint256 votes;
+        
     }
 
     /// @notice Possible states that a proposal may be in
